+++ ChairLoader/ChairLoader.vcxproj.filters
@@ -162,6 +162,21 @@
     <ClInclude Include="Prey\CryNetwork\ISerialize.h">
       <Filter>Prey\CryNetwork</Filter>
     </ClInclude>
+    <ClInclude Include="ChairloaderGUIConsole.h">
+      <Filter>Chairloader</Filter>
+    </ClInclude>
+    <ClInclude Include="ChairloaderGUIEntityManager.h">
+      <Filter>Chairloader</Filter>
+    </ClInclude>
+    <ClInclude Include="ChairloaderGUILog.h">
+      <Filter>Chairloader</Filter>
+    </ClInclude>
+    <ClInclude Include="ChairloaderGUIPlayerManager.h">
+      <Filter>Chairloader</Filter>
+    </ClInclude>
+    <ClInclude Include="GUIUtils.h">
+      <Filter>Chairloader</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="PreyDllObjects.cpp">
